+++ scripts/release/const.ts
@@ -1,0 +1,3 @@
+export const base = '../../packages/';
+export const registry = 'http://localhost:4873';
+// npm publish --registry http://localhost:4873
